@@ -3,37 +3,41 @@ package httplog
 import (
 	"context"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"net/http"
+	"os"
 	"strings"
+	"sync"
 	"time"
 
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/chi/v5/middleware"
-	"github.com/rs/zerolog"
-	"github.com/rs/zerolog/log"
+	"golang.org/x/exp/slog"
 )
 
-func NewLogger(serviceName string, opts ...Options) zerolog.Logger {
+func NewLogger(serviceName string, opts ...Options) *slog.Logger {
 	if len(opts) > 0 {
 		Configure(opts[0])
 	} else {
 		Configure(DefaultOptions)
 	}
-	logger := log.With().Str("service", strings.ToLower(serviceName))
+	logger := slog.With(slog.Attr{Key: "service", Value: slog.StringValue(strings.ToLower(serviceName))})
+	// logger := log.With().Str("service", strings.ToLower(serviceName))
 	if !DefaultOptions.Concise && len(DefaultOptions.Tags) > 0 {
-		logger = logger.Fields(map[string]interface{}{
-			"tags": DefaultOptions.Tags,
-		})
+		group := []slog.Attr{}
+		for k, v := range DefaultOptions.Tags {
+			group = append(group, slog.Attr{Key: k, Value: slog.StringValue(v)})
+		}
+		logger = logger.With(slog.Group("tags", group...))
 	}
-	return logger.Logger()
+	return logger
 }
 
 // RequestLogger is an http middleware to log http requests and responses.
 //
 // NOTE: for simplicity, RequestLogger automatically makes use of the chi RequestID and
 // Recoverer middleware.
-func RequestLogger(logger zerolog.Logger) func(next http.Handler) http.Handler {
+func RequestLogger(logger *slog.Logger) func(next http.Handler) http.Handler {
 	return chi.Chain(
 		middleware.RequestID,
 		Handler(logger),
@@ -41,10 +45,14 @@ func RequestLogger(logger zerolog.Logger) func(next http.Handler) http.Handler {
 	).Handler
 }
 
-func Handler(logger zerolog.Logger) func(next http.Handler) http.Handler {
-	var f middleware.LogFormatter = &requestLogger{logger}
+func Handler(logger *slog.Logger) func(next http.Handler) http.Handler {
+	var f middleware.LogFormatter = &requestLogger{*logger}
 	return func(next http.Handler) http.Handler {
 		fn := func(w http.ResponseWriter, r *http.Request) {
+			if rInCooldown(r) {
+				next.ServeHTTP(w, r)
+				return
+			}
 			entry := f.NewLogEntry(r)
 			ww := middleware.NewWrapResponseWriter(w, r.ProtoMajor)
 
@@ -55,7 +63,7 @@ func Handler(logger zerolog.Logger) func(next http.Handler) http.Handler {
 			defer func() {
 				var respBody []byte
 				if ww.Status() >= 400 {
-					respBody, _ = ioutil.ReadAll(buf)
+					respBody, _ = io.ReadAll(buf)
 				}
 				entry.Write(ww.Status(), ww.BytesWritten(), ww.Header(), time.Since(t1), respBody)
 			}()
@@ -67,21 +75,22 @@ func Handler(logger zerolog.Logger) func(next http.Handler) http.Handler {
 }
 
 type requestLogger struct {
-	Logger zerolog.Logger
+	Logger slog.Logger
 }
 
 func (l *requestLogger) NewLogEntry(r *http.Request) middleware.LogEntry {
 	entry := &RequestLoggerEntry{}
 	msg := fmt.Sprintf("Request: %s %s", r.Method, r.URL.Path)
-	entry.Logger = l.Logger.With().Fields(requestLogFields(r, true)).Logger()
+	entry.Logger = *l.Logger.With(requestLogFields(r, true))
 	if !DefaultOptions.Concise {
-		entry.Logger.Info().Fields(requestLogFields(r, DefaultOptions.Concise)).Msgf(msg)
+		entry.Logger = *l.Logger.With(requestLogFields(r, DefaultOptions.Concise))
+		entry.Logger.Info(msg)
 	}
 	return entry
 }
 
 type RequestLoggerEntry struct {
-	Logger zerolog.Logger
+	Logger slog.Logger
 	msg    string
 }
 
@@ -91,10 +100,10 @@ func (l *RequestLoggerEntry) Write(status, bytes int, header http.Header, elapse
 		msg = fmt.Sprintf("%s - %s", msg, l.msg)
 	}
 
-	responseLog := map[string]interface{}{
-		"status":  status,
-		"bytes":   bytes,
-		"elapsed": float64(elapsed.Nanoseconds()) / 1000000.0, // in milliseconds
+	responseLog := []slog.Attr{
+		{Key: "status", Value: slog.IntValue(status)},
+		{Key: "bytes", Value: slog.IntValue(bytes)},
+		{Key: "elapsed", Value: slog.Float64Value(float64(elapsed.Nanoseconds()) / 1000000.0)}, // in milliseconds
 	}
 
 	if !DefaultOptions.Concise {
@@ -102,16 +111,13 @@ func (l *RequestLoggerEntry) Write(status, bytes int, header http.Header, elapse
 		// the response body so we may inspect the log message sent back to the client.
 		if status >= 400 {
 			body, _ := extra.([]byte)
-			responseLog["body"] = string(body)
+			responseLog = append(responseLog, slog.Attr{Key: "body", Value: slog.StringValue(string(body))})
 		}
 		if len(header) > 0 {
-			responseLog["header"] = headerLogField(header)
+			responseLog = append(responseLog, slog.Group("header", headerLogField(header)...))
 		}
 	}
-
-	l.Logger.WithLevel(statusLevel(status)).Fields(map[string]interface{}{
-		"httpResponse": responseLog,
-	}).Msgf(msg)
+	l.Logger.With(slog.Group("httpResponse", responseLog...)).Log(statusLevel(status), msg)
 }
 
 func (l *RequestLoggerEntry) Panic(v interface{}, stack []byte) {
@@ -119,11 +125,12 @@ func (l *RequestLoggerEntry) Panic(v interface{}, stack []byte) {
 	if DefaultOptions.JSON {
 		stacktrace = string(stack)
 	}
-
-	l.Logger = l.Logger.With().
-		Str("stacktrace", stacktrace).
-		Str("panic", fmt.Sprintf("%+v", v)).
-		Logger()
+	l.Logger = *l.Logger.With(slog.Attr{Key: "stacktrace", Value: slog.StringValue(stacktrace)},
+		slog.Attr{Key: "panic", Value: slog.StringValue(fmt.Sprintf("%+v", v))})
+	// l.Logger = l.Logger.With().
+	// 	Str("stacktrace", stacktrace).
+	// 	Str("panic", fmt.Sprintf("%+v", v)).
+	// 	Logger()
 
 	l.msg = fmt.Sprintf("%+v", v)
 
@@ -132,60 +139,102 @@ func (l *RequestLoggerEntry) Panic(v interface{}, stack []byte) {
 	}
 }
 
-func requestLogFields(r *http.Request, concise bool) map[string]interface{} {
+var coolDownMu sync.RWMutex
+var coolDowns = map[string]time.Time{}
+
+func rInCooldown(r *http.Request) bool {
+	routePath := r.URL.EscapedPath()
+	if routePath == "" {
+		routePath = "/"
+	}
+	if !inArray(DefaultOptions.QuietDownRoutes, routePath) {
+		return false
+	}
+	coolDownMu.RLock()
+	coolDownTime, ok := coolDowns[routePath]
+	coolDownMu.RUnlock()
+	if ok {
+		if time.Since(coolDownTime) < DefaultOptions.QuietDownPeriod {
+			return true
+		}
+	}
+	coolDownMu.Lock()
+	defer coolDownMu.Unlock()
+	coolDowns[routePath] = time.Now().Add(DefaultOptions.QuietDownPeriod)
+	return false
+}
+
+func inArray(arr []string, val string) bool {
+	for _, v := range arr {
+		if v == val {
+			return true
+		}
+	}
+	return false
+}
+
+func requestLogFields(r *http.Request, concise bool) slog.Attr {
 	scheme := "http"
 	if r.TLS != nil {
 		scheme = "https"
 	}
 	requestURL := fmt.Sprintf("%s://%s%s", scheme, r.Host, r.RequestURI)
 
-	requestFields := map[string]interface{}{
-		"requestURL":    requestURL,
-		"requestMethod": r.Method,
-		"requestPath":   r.URL.Path,
-		"remoteIP":      r.RemoteAddr,
-		"proto":         r.Proto,
+	requestFields := []slog.Attr{
+		{Key: "requestURL", Value: slog.StringValue(requestURL)},
+		{Key: "requestMethod", Value: slog.StringValue(r.Method)},
+		{Key: "requestPath", Value: slog.StringValue(r.URL.Path)},
+		{Key: "remoteIP", Value: slog.StringValue(r.RemoteAddr)},
+		{Key: "proto", Value: slog.StringValue(r.Proto)},
 	}
 	if reqID := middleware.GetReqID(r.Context()); reqID != "" {
-		requestFields["requestID"] = reqID
+		requestFields = append(requestFields, slog.Attr{Key: "requestID", Value: slog.StringValue(reqID)})
+		// requestFields["requestID"] = reqID
 	}
 
 	if concise {
-		return map[string]interface{}{
-			"httpRequest": requestFields,
-		}
+		return slog.Group("httpRequest", requestFields...)
 	}
 
-	requestFields["scheme"] = scheme
-
+	// requestFields["scheme"] = scheme
+	requestFields = append(requestFields, slog.Attr{Key: "scheme", Value: slog.StringValue(scheme)})
 	if len(r.Header) > 0 {
-		requestFields["header"] = headerLogField(r.Header)
+		// requestFields["header"] = headerLogField(r.Header)
+		requestFields = append(requestFields,
+			slog.Attr{Key: "header",
+				Value: slog.GroupValue(headerLogField(r.Header)...)})
 	}
 
-	return map[string]interface{}{
-		"httpRequest": requestFields,
-	}
+	return slog.Group("httpRequest", requestFields...)
 }
 
-func headerLogField(header http.Header) map[string]string {
-	headerField := map[string]string{}
+func headerLogField(header http.Header) []slog.Attr {
+	headerField := []slog.Attr{}
 	for k, v := range header {
 		k = strings.ToLower(k)
 		switch {
 		case len(v) == 0:
 			continue
 		case len(v) == 1:
-			headerField[k] = v[0]
+			headerField = append(headerField, slog.Attr{Key: k, Value: slog.StringValue(v[0])})
 		default:
-			headerField[k] = fmt.Sprintf("[%s]", strings.Join(v, "], ["))
+			headerField = append(headerField, slog.Attr{Key: k,
+				Value: slog.StringValue(fmt.Sprintf("[%s]", strings.Join(v, "], [")))})
+			// headerField = fmt.Sprintf("[%s]", strings.Join(v, "], ["))
 		}
 		if k == "authorization" || k == "cookie" || k == "set-cookie" {
-			headerField[k] = "***"
+			headerField[len(headerField)] = slog.Attr{
+				Key:   k,
+				Value: slog.StringValue("***"),
+			}
 		}
 
 		for _, skip := range DefaultOptions.SkipHeaders {
 			if k == skip {
-				headerField[k] = "***"
+				headerField[len(headerField)] = slog.Attr{
+					Key:   k,
+					Value: slog.StringValue("***"),
+				}
 				break
 			}
 		}
@@ -193,18 +242,18 @@ func headerLogField(header http.Header) map[string]string {
 	return headerField
 }
 
-func statusLevel(status int) zerolog.Level {
+func statusLevel(status int) slog.Level {
 	switch {
 	case status <= 0:
-		return zerolog.WarnLevel
+		return slog.LevelWarn
 	case status < 400: // for codes in 100s, 200s, 300s
-		return zerolog.InfoLevel
+		return slog.LevelInfo
 	case status >= 400 && status < 500:
-		return zerolog.WarnLevel
+		return slog.LevelWarn
 	case status >= 500:
-		return zerolog.ErrorLevel
+		return slog.LevelError
 	default:
-		return zerolog.InfoLevel
+		return slog.LevelInfo
 	}
 }
 
@@ -230,10 +279,17 @@ func statusLabel(status int) string {
 // passes through the handler chain, which at any point can be logged
 // with a call to .Print(), .Info(), etc.
 
-func LogEntry(ctx context.Context) zerolog.Logger {
+func LogEntry(ctx context.Context) slog.Logger {
 	entry, ok := ctx.Value(middleware.LogEntryCtxKey).(*RequestLoggerEntry)
 	if !ok || entry == nil {
-		return zerolog.Nop()
+		handlerOpts := &slog.HandlerOptions{
+			AddSource: true,
+			// LevelError+1 will be higher than all levels
+			// hence logs would be skipped
+			Level: slog.LevelError + 1,
+			// ReplaceAttr: func(attr slog.Attr) slog.Attr ,
+		}
+		return *slog.New(handlerOpts.NewTextHandler(os.Stdout))
 	} else {
 		return entry.Logger
 	}
@@ -241,12 +297,18 @@ func LogEntry(ctx context.Context) zerolog.Logger {
 
 func LogEntrySetField(ctx context.Context, key, value string) {
 	if entry, ok := ctx.Value(middleware.LogEntryCtxKey).(*RequestLoggerEntry); ok {
-		entry.Logger = entry.Logger.With().Str(key, value).Logger()
+		entry.Logger = *entry.Logger.With(slog.Attr{Key: key, Value: slog.StringValue(value)})
 	}
 }
 
 func LogEntrySetFields(ctx context.Context, fields map[string]interface{}) {
 	if entry, ok := ctx.Value(middleware.LogEntryCtxKey).(*RequestLoggerEntry); ok {
-		entry.Logger = entry.Logger.With().Fields(fields).Logger()
+		attrs := make([]slog.Attr, len(fields))
+		i := 0
+		for k, v := range fields {
+			attrs[i] = slog.Attr{Key: k, Value: slog.AnyValue(v)}
+			i++
+		}
+		entry.Logger = *entry.Logger.With(attrs)
 	}
 }
