@@ -1,22 +1,33 @@
 package main
 
 import (
+	"errors"
 	"net/http"
+	"time"
 
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/chi/v5/middleware"
 	"github.com/go-chi/httplog"
+	"golang.org/x/exp/slog"
 )
 
 func main() {
 	// Logger
 	logger := httplog.NewLogger("httplog-example", httplog.Options{
-		// JSON: true,
-		Concise: true,
-		// Tags: map[string]string{
-		// 	"version": "v1.0-81aa4244d9fc8076a",
-		// 	"env":     "dev",
-		// },
+		JSON:            false,
+		LogLevel:        slog.LevelDebug,
+		Concise:         true,
+		TimeFieldFormat: time.RFC850,
+		Tags: map[string]string{
+			"version": "v1.0-81aa4244d9fc8076a",
+			"env":     "dev",
+		},
+		QuietDownRoutes: []string{
+			"/",
+			"/ping",
+		},
+		QuietDownPeriod: 10 * time.Second,
+		// SourceFieldName: "source",
 	})
 
 	// Service
@@ -35,23 +46,23 @@ func main() {
 	r.Get("/info", func(w http.ResponseWriter, r *http.Request) {
 		oplog := httplog.LogEntry(r.Context())
 		w.Header().Add("Content-Type", "text/plain")
-		oplog.Info().Msg("info here")
+		oplog.Info("info here")
 		w.Write([]byte("info here"))
 	})
 
 	r.Get("/warn", func(w http.ResponseWriter, r *http.Request) {
 		oplog := httplog.LogEntry(r.Context())
-		oplog.Warn().Msg("warn here")
+		oplog.Warn("warn here")
 		w.WriteHeader(400)
 		w.Write([]byte("warn here"))
 	})
 
 	r.Get("/err", func(w http.ResponseWriter, r *http.Request) {
 		oplog := httplog.LogEntry(r.Context())
-		oplog.Error().Msg("err here")
+		oplog.Error("msg here", errors.New("err here"))
 		w.WriteHeader(500)
 		w.Write([]byte("err here"))
 	})
 
-	http.ListenAndServe(":5555", r)
+	http.ListenAndServe(":8000", r)
 }
